+++ opencxl/cxl/mmio/component_register/memcache_register/__init__.py
@@ -10,7 +10,7 @@ from opencxl.cxl.mmio.component_register.memcache_register.cache_route_table imp
     CxlCacheIdRTCapabilityStructure4N,
     CacheRouteTableCapabilityStructureOptions4N,
 )
-from opencxl.cxl.mmio.component_register.memcache_register.cache_id_decoder_capabiility import (
+from opencxl.cxl.mmio.component_register.memcache_register.cache_id_decoder_capability import (
     CxlCacheIdDecoderCapabilityStructure,
     CxlCacheIdDecoderCapabilityStructureOptions,
 )
